The player's own controls, and a release for the filter
Ten controls that sit on top of whatever programme is loaded, every one an offset from
what the disk says and every one zero by default, so a programme plays as recorded until
something is moved. An absolute control would flatten a programme's keygroups to a single
value the moment it was touched; an offset moves the shape and keeps it.

  cutoff 74   amount 70
  VCA     attack 73   decay 75   sustain 79   release 72
  VCF     attack 102  decay 103  sustain 104  release 105

The controllers MOVE the parameters rather than running alongside them, so a knob, an
automation lane and a keyboard are three ways at one control rather than three that
disagree. The envelopes are shapes with draggable corners rather than eight knobs, since
the thing being adjusted is a shape.

THE FILTER HAD NO RELEASE

Not here and not in the C#: the fourth VCF byte was read off the disk, carried in the
patch and never used. It has one now - the envelope falls back to the keygroup's own
cutoff as the note dies, from wherever it had reached rather than from the sustain.

That needed the filter its own clock, and finding out why is the more useful half. It was
being driven by the AMPLITUDE envelope's stage timer, which is reset to zero every time
that envelope changes stage - so a programme with a slow amplitude decay restarted its
filter sweep when the decay ended. Unnoticed because most of the library has an instant
attack and decay, which resets a clock that has barely started. Two renders differing only
in vcaDecay must now give the same filter movement, and a check holds that.

MEASURED ALONG THE WAY

cutoffHz and envSeconds take a double now. 127 controller steps across a 198-step range
land between the panel's own, and truncating to int made the controls step - audibly, for
envelope times, where the curve is exponential. Whole numbers are unchanged, so the
conformance figures and Reference.h are untouched.

ControlBlock stays at 32. A sixth-order cascade retuned from 17760 Hz to 311 Hz in one
step rings at 17.85 against a signal of 1.0 at a block of 64, and 1.24 at 32, with no
further gain below that - so it was already past the cliff, and going finer would have
bought nothing for four times the modulator work.

+++ Plugin/Source/S950/Cal.h
@@ -129,12 +129,18 @@ namespace s950::cal
      *
      * The filter is also the reconstruction filter, so the top of its travel moves with the
      * rate the audio comes out at rather than being a fixed frequency.
+     *
+     * `stored` is a double rather than the panel's integer because the plugin's cutoff trim
+     * slides between the panel's steps: 127 MIDI values across a 198-step range lands between
+     * them constantly, and rounding would step the filter audibly. For a whole number this
+     * returns exactly what it always did - the value was widened to double on entry anyway -
+     * so the conformance numbers are untouched.
      */
-    inline double cutoffHz (int stored, double rate)
+    inline double cutoffHz (double stored, double rate)
     {
         const double ceiling = MaxRatio * (rate > 0 ? rate : 48000.0);
         const double floor   = std::min (FloorHz, ceiling);
-        const double v       = clamp (stored, 0, 99);
+        const double v       = clamp (stored, 0.0, 99.0);
 
         double hz = curveAt (CurveCount - 1, ceiling);
 
@@ -155,10 +161,17 @@ namespace s950::cal
         return clamp (hz, floor, ceiling);
     }
 
-    /// A stored 0..99 envelope time, in seconds.
-    inline double envSeconds (int stored)
+    /*
+     * A stored 0..99 envelope time, in seconds.
+     *
+     * A double for the same reason cutoffHz takes one: the plugin's envelope trims land
+     * between the panel's steps, and this curve is exponential - a whole unit is about a
+     * tenth of the time either way, which steps audibly. Truncating to int was also a silent
+     * narrowing that the compiler was right to complain about. Whole numbers are unchanged.
+     */
+    inline double envSeconds (double stored)
     {
-        const double v = clamp (stored, 0, 99) / 99.0;
+        const double v = clamp (stored, 0.0, 99.0) / 99.0;
         return (EnvMinMs * std::pow (EnvMaxMs / EnvMinMs, v)) / 1000.0;
     }
